added another stage for plugins
there are now 3 stages of cxx-plugins. each plugin needs a plugin_name/plugin.rb with a cxx_plugin block.
depending on the stage the block has to take a different number of arguments.

1. at first plugins with no argument are loaded (before the project-files are evaluated)
2. directly after this all toolchains are initialized (the toolchains should have no arguments to the block)
3. after all building blocks the rest of the plugin with 3 arguments to the block are loaded

TODO .. join the first two stages .. both are plugins with no arguments and this is something which distinguishes them from the plugins with 3 arguments

Change-Id: Ib54e6ca2683694ed4ada9b0775b4c7bb650c375c